// src/Models/HistoricoDois.model.ts
export interface historicoDoisModel {
   id: number;
    cliente: string;
    cpf: string;
    quantidade: number;
   data_hora: Date;
   operacao: string;
   item: string
    
    
}
export interface historicoDoisModelPost {
    //id: number;
    cliente: string;
    cpf: string;
    quantidade: number;
    operacao: string;
    item: string
   //data_hora: Date;
    
    
}